# contracts/openapi-internal.yaml
openapi: 3.0.3
info:
  title: ProjetoX Internal API
  version: 0.1.0
servers:
  - url: http://localhost:8080
paths:
  /calendar/events:
    post:
      summary: Cria evento
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateEventRequest'
      responses:
        '201':
          description: Evento criado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EventResponse'
    get:
      summary: Lista eventos por janela
      parameters:
        - in: query
          name: from
          required: true
          schema:
            type: string
            format: date-time
        - in: query
          name: to
          required: true
          schema:
            type: string
            format: date-time
      responses:
        '200':
          description: Lista de eventos
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/EventResponse'
  /calendar/events/{id}:
    delete:
      summary: Cancela evento
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
      responses:
        '204':
          description: Evento removido
        '404':
          description: Não encontrado
components:
  schemas:
    CreateEventRequest:
      type: object
      required: [title, start, end, timezone]
      properties:
        title:
          type: string
        start:
          type: string
          format: date-time
        end:
          type: string
          format: date-time
        timezone:
          type: string
    EventResponse:
      type: object
      properties:
        id:
          type: string
        title:
          type: string
        start:
          type: string
          format: date-time
        end:
          type: string
          format: date-time
        timezone:
          type: string