# services/calendar-core-java/src/main/java/com/projetox/calendar/api/dto/CreateEventRequest.java
package com.projetox.calendar.api.dto;

import jakarta.validation.constraints.NotBlank;

public record CreateEventRequest(
        @NotBlank String title,
        @NotBlank String start,
        @NotBlank String end,
        @NotBlank String timezone
) {
}